Add UserServiceTest
