+++ src/main/java/com/jjcsa/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.jjcsa.repository;
 
 import com.jjcsa.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, UUID> {
+public interface UserRepository extends JpaRepository<User, UUID> {
 
     User findUserById(UUID id);
